+++ client/openastroara_client/test/screens/tabs/sequencer_tab_test.dart
@@ -59,6 +59,8 @@ class _FakeClient implements SequenceClient {
   @override
   Future<String> resume(String id) async => 'op';
   @override
+  Future<String> skipCurrent(String id) async => 'op';
+  @override
   Future<String> abort(String id) async => 'op';
   @override
   Future<String> stop(String id) async => 'op';